# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyTravel API",
        "description": "Documentation for EasyTravel project.",
        "contact": {
            "email": "21732184@live.uem.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/airports/coordinates": {
            "get": {
                "tags": [
                    "airports"
                ],
                "summary": "Get all airports coordinates with URL",
                "description": "Returns all the airports coordinates and basic info.",
                "operationId": "getAirportsCoordinates",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All coordinates",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "airport_id": 216,
                                            "airport_name": "Cootamundra Airport",
                                            "airport_country": "Australia",
                                            "airport_city": "",
                                            "airport_lon": 148.028,
                                            "airport_lat": -34.623901
                                        },
                                        {
                                            "airport_id": 241,
                                            "airport_name": "Adelaide International Airport",
                                            "airport_country": "Australia",
                                            "airport_city": "Adelaide",
                                            "airport_lon": 138.53101,
                                            "airport_lat": -34.945
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/cities/cities": {
            "get": {
                "tags": [
                    "cities"
                ],
                "summary": "Get all cities",
                "description": "Returns all the cities  of the database.",
                "operationId": "getCities",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All models",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": 1,
                                            "name": "Bayman"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Camp Bastion"
                                        },
                                        {
                                            "id": 30,
                                            "name": "Annaba"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/cities/top": {
            "get": {
                "tags": [
                    "comments",
                    "cities"
                ],
                "summary": "Top 4 cities by user sentients",
                "description": "Returns the top 4 destinations in base of the sentiment average",
                "operationId": "getAirportsCoordinates",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All coordinates",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "city": "Madrid",
                                            "grade": "9.463677",
                                            "sentiment": 0.633517016156253
                                        },
                                        {
                                            "city": "Burgos",
                                            "grade": "8.724951",
                                            "sentiment": 0.6041372706116136
                                        },
                                        {
                                            "city": "Pisa",
                                            "grade": "9.207483",
                                            "sentiment": 0.6028689742527156
                                        },
                                        {
                                            "city": "Paris",
                                            "grade": "9.041719",
                                            "sentiment": 0.5961088823466875
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}